+++ packages/analytics/src/index.ts
@@ -4,8 +4,6 @@ export { getVisitor, setVisitor } from './visitor/index';
 export { sendFeedback } from './feedback/index';
 export { createLink, getLink, type Link } from './link/index';
 export {
-  Platform,
-  Environment,
   ALL_PLATFORMS,
   ALL_ENVIRONMENTS,
   createTrackEventSchemaV1,
@@ -20,8 +18,10 @@ export {
 export { stripeMinorUnits } from './utils/stripe';
 
 export type {
-  TrackProperties,
+  Platform,
+  Environment,
   TrackTags,
+  TrackProperties,
   AllowedPropertyValues,
   UserProvidedData,
 } from './track/types';