+++ packages/analytics/src/setup/index.ts
@@ -1,4 +1,4 @@
-import type { ThirdPartyTracker, TrackTags } from '../track/types';
+import type { Environment, Platform, ThirdPartyTracker, TrackTags } from '../track/types';
 import type { ThirdPartyUserSetter, Visitor } from '../visitor/types';
 
 export interface Storage {
@@ -10,6 +10,8 @@ export interface Options {
   release: string;
   storage: Storage;
   endpoint: string;
+  platform: Platform;
+  environment: Environment;
   getTags: () => TrackTags | Promise<TrackTags>;
   getDeviceId: () => string | Promise<string>;
   getHeaders?: () => Record<string, string> | Promise<Record<string, string>>;
@@ -21,6 +23,8 @@ interface Config {
   release: string;
   endpoint: string;
   storage: Storage;
+  platform: Platform;
+  environment: Environment;
   getTags: () => TrackTags | Promise<TrackTags>;
   getDeviceId: () => string | Promise<string>;
   getHeaders: () => Record<string, string> | Promise<Record<string, string>>;
@@ -42,6 +46,8 @@ export const config: Config = {
   endpoint: '',
   release: '0.0.0',
   storage: null!,
+  platform: null!,
+  environment: null!,
   getTags: null!,
   getDeviceId: null!,
   getHeaders: null!,
@@ -52,6 +58,8 @@ export const config: Config = {
 export function setupAnalytics(init: Options) {
   config.release = init.release;
   config.storage = init.storage;
+  config.platform = init.platform;
+  config.environment = init.environment;
   config.endpoint = init.endpoint.endsWith('/') ? init.endpoint.slice(0, -1) : init.endpoint;
   config.getTags = init.getTags;
   config.getDeviceId = init.getDeviceId;