+++ packages/analytics/src/visitor/types.ts
@@ -1,16 +1,14 @@
 import type { UserProperties, UserProvidedData } from '../track/gtag';
+import type { Environment, Platform } from '../track/types';
 
 export type VisitorProperties = Record<Lowercase<string>, string | number | boolean | null>;
 
 export interface Visitor {
   id: string;
   device_id: string;
+  platform: Platform;
+  environment: Environment;
   properties: VisitorProperties;
-}
-
-export interface CreateVisitorDTO {
-  device_id: string;
-  properties?: VisitorProperties;
 }
 
 export interface UpdateVisitorDTO {